fastapi/requirements.txt
fastapi
pydantic
numpy
uvicorn
pandas
matplotlib
azure-storage-blob 
azure-identity